pyproject.toml
[tool.poetry]
name = "text-generation-apps"
version = "0.1.0"
description = "Text generation app"
authors = ["hungpm372 <hungpm372@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.35.13"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
